requirements.txt
pillow
python-dotenv
python-telegram-bot
telethon